requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
python-dotenv